refactor header and other related components and types to be shared via ui package
