+++ apps/vbpay/components/header.tsx
@@ -1,27 +1,24 @@
 "use client";
 
-import { Fragment } from "react";
+import { startTransition } from "react";
 import { useParams, usePathname } from "next/navigation";
+import { updateUserSelectionSlugAction } from "@/actions/update-user-selection-slug-action";
 import { getPageTitle } from "@/utils/get-page-title";
+import { APP_NAME } from "@/values/app-name";
+import { signOut } from "aws-amplify/auth";
+import { useAction } from "next-safe-action/hooks";
+import { parseAsString, useQueryState } from "nuqs";
 
-import {
-  Breadcrumb,
-  BreadcrumbItem,
-  BreadcrumbLink,
-  BreadcrumbList,
-  BreadcrumbPage,
-  BreadcrumbSeparator,
-} from "@workspace/ui/components/breadcrumb";
-import { Separator } from "@workspace/ui/components/separator";
-import { SidebarTrigger } from "@workspace/ui/components/sidebar";
+import { AppHeader } from "@workspace/ui/components/common/app-header";
+import { ErrorDialog } from "@workspace/ui/components/error-dialog";
+import { clearSidebarState } from "@workspace/ui/components/main-sidebar/main-sidebar-cookies";
+import { useErrorDialog } from "@workspace/ui/hooks/use-error-dialog";
+import { BreadcrumbItemType } from "@workspace/ui/types/breadcrumb-item";
+import { UserSelectionData } from "@workspace/ui/types/user-selection-data";
 
-import { BreadcrumbItemType } from "@/types/breadcrumb-item";
-import { UserSelectionData } from "@/types/user-selection-data";
-import { SignOutButton } from "@/components/signout-button";
+import { Icons } from "@/components/icons";
 
-import { UserSelectionCombo } from "./user-selection-combo";
-
-type props = {
+type Props = {
   breadcrumbs?: BreadcrumbItemType[];
   overrideTitle?: string;
   userSelectionData?: UserSelectionData;
@@ -31,59 +28,70 @@ export function Header({
   breadcrumbs,
   overrideTitle,
   userSelectionData,
-}: props) {
+}: Props) {
   const pathname = usePathname();
   const params = useParams();
-  const title =
-    overrideTitle ?? getPageTitle(pathname, params.slug as string | undefined);
+  const [pId, setPId] = useQueryState(
+    "pId",
+    parseAsString
+      .withDefault("")
+      .withOptions({ startTransition, shallow: false }),
+  );
+  const {
+    openErrorDialog,
+    isErrorDialogOpen,
+    errorMsg,
+    errorTitle,
+    closeErrorDialog,
+  } = useErrorDialog({});
+
+  const { execute } = useAction(updateUserSelectionSlugAction, {
+    onError: ({ error }) => {
+      openErrorDialog(
+        "Error",
+        error.validationErrors
+          ? "Invalid inputs. Please double check the data and try again. If the problem persists please contact support."
+          : error.serverError
+            ? error.serverError
+            : (error as string),
+      );
+    },
+  });
 
   return (
-    <header className="border-b p-5">
-      <div className="container mx-auto flex items-center justify-between gap-4">
-        <div className="flex justify-between">
-          <div>
-            <div className="flex items-center gap-2">
-              <SidebarTrigger className="-ml-1" />
-              <Separator orientation="vertical" className="mr-2 h-4" />
-              <h1 className="text-3xl font-bold" data-testid="header-title">
-                {title}
-              </h1>
-            </div>
-            <Breadcrumb className="mt-2">
-              <BreadcrumbList>
-                {breadcrumbs?.map((item, index) => (
-                  <Fragment key={item.title}>
-                    <BreadcrumbItem
-                      className={
-                        index < breadcrumbs.length - 1 ? "hidden md:block" : ""
-                      }
-                    >
-                      {item.href ? (
-                        <BreadcrumbLink href={item.href}>
-                          {item.title}
-                        </BreadcrumbLink>
-                      ) : (
-                        <BreadcrumbPage>{item.title}</BreadcrumbPage>
-                      )}
-                    </BreadcrumbItem>
-                    {index < breadcrumbs.length - 1 && (
-                      <BreadcrumbSeparator className="hidden md:block" />
-                    )}
-                  </Fragment>
-                ))}
-              </BreadcrumbList>
-            </Breadcrumb>
-          </div>
-        </div>
-        {userSelectionData && (
-          <UserSelectionCombo
-            slug={userSelectionData.slug}
-            comboItems={userSelectionData.comboItems}
-            defaultLock={userSelectionData.defaultLock}
-          />
-        )}
-        <SignOutButton />
-      </div>
-    </header>
+    <>
+      {isErrorDialogOpen && (
+        <ErrorDialog
+          open={isErrorDialogOpen}
+          onOpenChange={closeErrorDialog}
+          description={errorMsg}
+          title={errorTitle}
+        />
+      )}
+      <AppHeader
+        breadcrumbs={breadcrumbs}
+        overrideTitle={overrideTitle}
+        userSelectionData={userSelectionData}
+        getPageTitle={getPageTitle}
+        signOut={signOut}
+        onSignOut={() => clearSidebarState(APP_NAME)}
+        pathname={pathname}
+        slug={params.slug as string | undefined}
+        userSelectionInstructText="Select Payer..."
+        userSelectionNotFoundText="Payer Not Found"
+        userSelectionSearchText="Search..."
+        userSelectionUpdate={execute}
+        userSelectionValue={pId ?? ""}
+        onUserSelectionValueChange={setPId}
+        userSelectionIcons={{
+          chevronsUpDown: (
+            <Icons.chevronsUpDown className="ml-2 size-4 shrink-0 opacity-50" />
+          ),
+          check: <Icons.check className="size-4" />,
+          lock: <Icons.lock className="size-5" />,
+          unlock: <Icons.unlock className="size-5" />,
+        }}
+      />
+    </>
   );
 }
