--- apps/vbpay/types/breadcrumb-item.ts
@@ -1,4 +1,0 @@
-export type BreadcrumbItemType = {
-  title: string;
-  href?: string;
-};
