--- apps/vbpay/types/user-selection-data.ts
@@ -1,7 +1,0 @@
-import { ComboItem } from "@workspace/ui/types/combo-item";
-
-export type UserSelectionData = {
-  defaultLock: boolean;
-  slug: string | undefined;
-  comboItems: ComboItem[];
-};
