+++ packages/ui/src/components/common/signout-button.tsx
@@ -4,29 +4,29 @@ import { LogOut } from "lucide-react";
 
 import { Button } from "@workspace/ui/components/button";
 import { ConfirmationDialog } from "@workspace/ui/components/common/confirmation-dialog";
+import { ErrorDialog } from "@workspace/ui/components/error-dialog";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipTrigger,
+} from "@workspace/ui/components/tooltip";
 import { useConfirmationDialog } from "@workspace/ui/hooks/use-confirmation-dialog";
+import { useErrorDialog } from "@workspace/ui/hooks/use-error-dialog";
 
 type props = {
   signOut: () => Promise<void>;
-  variant?:
-    | "default"
-    | "destructive"
-    | "outline"
-    | "secondary"
-    | "ghost"
-    | "link";
-  size?: "default" | "sm" | "lg" | "icon";
-  className?: string;
-  children?: React.ReactNode;
+  onSignOut?: () => void;
 };
 
-export function SignoutButton({
-  signOut,
-  variant = "ghost",
-  size = "default",
-  className,
-  children,
-}: props) {
+export function SignoutButton({ signOut, onSignOut }: props) {
+  const {
+    openErrorDialog,
+    isErrorDialogOpen,
+    errorMsg,
+    errorTitle,
+    closeErrorDialog,
+  } = useErrorDialog({});
+
   const {
     confirm,
     cancel,
@@ -35,7 +35,17 @@ export function SignoutButton({
     confDialogTitle,
     confDialogMsg,
   } = useConfirmationDialog({
-    onConfirmAsync: signOut,
+    onConfirmAsync: async () => {
+      try {
+        if (onSignOut) {
+          onSignOut();
+        }
+        await signOut();
+      } catch (err) {
+        const error = err as Error;
+        openErrorDialog("Error Signing Out", error.message);
+      }
+    },
     onConfirmAfterAsync: () => {
       window.location.reload();
     },
@@ -52,17 +62,31 @@ export function SignoutButton({
           title={confDialogTitle}
         />
       )}
-      <Button
-        variant={variant}
-        size={size}
-        className={className}
-        onClick={() =>
-          openConfDialog("Sign Out", "Are you sure you want to sign out?")
-        }
-      >
-        <LogOut className="mr-2 h-4 w-4" />
-        {children || "Sign out"}
-      </Button>
+      {isErrorDialogOpen && (
+        <ErrorDialog
+          open={isErrorDialogOpen}
+          title={errorTitle}
+          description={errorMsg}
+          onOpenChange={closeErrorDialog}
+        />
+      )}
+      <Tooltip>
+        <TooltipTrigger asChild>
+          <Button
+            onClick={() =>
+              openConfDialog("Sign Out", "Are you sure you want to sign out?")
+            }
+            variant="ghost"
+            size="icon"
+            data-testid="sign-out-button"
+          >
+            <LogOut />
+          </Button>
+        </TooltipTrigger>
+        <TooltipContent data-testid="sign-out-button-tooltip-content">
+          Sign Out
+        </TooltipContent>
+      </Tooltip>
     </>
   );
 }
